Revert 9be86deeba7761772addee0c2f83e1c9583b2d5f
The commit 9be86deeba7761772addee0c2f83e1c9583b2d5f introduced
a hack for __builtin_expect() but that served only to hide the
real issue, which is that the expression pattern was not being
fetched correctly.  Its not at 5 and not 3.  Fix this, and the
original bug goes away.