Restore lost PIC code
This version of gcc seems not to have contained the latest patches
for the PIC code, needed for compiling glibc. This patches were
in the December 1999 cygnus CVS code repo, but not in gnu. The loss
of these patches was probably due to the egcs infighting that was
happening at that time.