Avoid crazy LA r9,-9223372036854775808(0,0)
Long story:  without this fix, I get
.L29:
   SLR   r8,r8
   LA r9,-9223372036854775808(0,0)
   CLM   r12,1,=XL1'00'
   BNE   .L31
   L  r8,=F'4294967295'
   SRDA  r8,32
where -9223372036854775808 is hex 0x800000000000000
That is minus INT_MAX. Printed this way because... that is
what INTVAL(XV) actually is. Casting this to an unsigned int
gives zero, and the code this a zero fits into 12 bits. Ooops.

This gets triggered on some widechar_t cod which is explicitly
fiddling with long long ints. Oh well.

Originally seen on the movdi pattern in i370.md at line 844
but applying conservative fix everywhere.