Remove emergency failsafe base reg reload.
The sequence of patches from the last few days finally fix the
root cause of the missing base reg reloads. The emergency-failsafe
is no longer needed.

None of this makes much of a difference, because these situations
are rare, and any savings/benefit is minimal and possible none at
all. But "doing the right thing", at least conceptually, is worth
something, just so that hacks don't accumulate.